Fix bug in crossSection_plugin and add warning in lasagna_helperFunctions
crossSection was failing if the mouse is above the first plane in the bottom
left panel (XZ plot). Check if Y is negative and return if it is fixed that.

crossSection also failed when the mouse was in the right bottom panel (empty
panel right of the XZ plot and below the YZ plot). Add a hasattr check in
lasagna_helperFunctions fixed that. Now it prints a warning.